% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adaptiveWeightsCpp}
\alias{adaptiveWeightsCpp}
\title{Calculate Multitaper Adaptive Weights}
\usage{
adaptiveWeightsCpp(eigenSpec, ev, k, nFFT)
}
\arguments{
\item{eigenSpec}{A \code{matrix} with columns containing the eigenspectra}

\item{ev}{A \code{vector} with the eigenvalues of the Slepian tapers.}

\item{k}{A \code{numeric} integer representing the number of tapers to use.}

\item{nFFT}{A \code{numeric} integer representing the total number of frequency bins 
that were used to estimate the spectra.}
}
\value{
A \code{matrix} containing the weights for each eigenspectra at each 
frequency.
}
\description{
Used to calculate the adaptive weights on eigenspectra using a bound on the 
broadband bias.
}
\details{
When estimating the variance, this function will overestimate because 
it includes the zero and nyquist frequency twice (should only be once).
}
