% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{blockedEigenCoef}
\alias{blockedEigenCoef}
\title{Calculates the eigencoefficents of a blocked time series}
\usage{
blockedEigenCoef(x, deltat = 1, nw, k, nFFT, numSections = length(x),
  adaptiveWeighting = TRUE, returnWeights = FALSE, idx = NULL)
}
\arguments{
\item{x}{A \code{list} of data.frames as returned by \link{sectionData}.}

\item{deltat}{A \code{numeric} indicating the sample rate.}

\item{nw}{A \code{numeric} indicating the time bandwidth parameter for estimating the 
Slepian data tapers.}

\item{k}{A \code{numeric} indicating the number of tapers to use - should be approximately
floor(2*nw - 1) and no larger than floor(2*nw).}

\item{nFFT}{A \code{numeric} indicating the number of frequency bins to use (i.e. setting 
the zeropadding amount).}

\item{numSections}{A \code{numeric} indicating the number of elements in the \code{list} x.}

\item{adaptiveWeighting}{A \code{logical} indicating whether the eigencoefficients should 
be multiplied by their adaptive weights before performing the regression.}
}
\value{
A \code{list} of lists.  You probably want to call \link{stackEigenByFreq} 
after this to get a more usable format.
}
\description{
Calculates the eigencoefficients (tapered Fourier transforms) of the provided (pre-blocked)
series.
}
\details{
Takes a \code{list} of data.frames and calculates the tapered Fourier transform 
of each column of each data.frame.  \link{spec.mtm} will calculate the adaptive weights 
to be used (bias protection) and these can also (and probably should be) used during 
coherence and transfer function estimation.
}
