% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{coherence}
\alias{coherence}
\title{Calculates the coherence between two series}
\usage{
coherence(x, y, blockSize = length(x), overlap = 0, deltat = 1, nw = 4,
  k = 7, nFFT = NULL, freqRange = NULL, maxFreqOffset = NULL,
  standardize = TRUE, prewhiten = TRUE, removePeriodic = TRUE,
  sigCutoff = NULL)
}
\arguments{
\item{x}{A \code{numeric} vector representing the first time series.}

\item{y}{A \code{numeric} vector containing the response series.}

\item{blockSize}{A \code{numeric} indicating the block sizes into which the 
input and response series will be partitioned.}

\item{overlap}{A \code{numeric} between 0 and 1 indicating how much overlap should exist 
between adjacent blocks.}

\item{deltat}{A \code{numeric} indicating the sample rate.}

\item{nw}{A \code{numeric} indicating the time bandwidth parameter for estimating the 
Slepian data tapers.}

\item{k}{A \code{numeric} indicating the number of tapers to use - should be approximately
floor(2*nw - 1) and no larger than floor(2*nw).}

\item{nFFT}{A \code{numeric} indicating the number of frequency bins to use (i.e. setting 
the zeropadding amount).}
}
\description{
Estimates the frequency domain coherence using the multitaper method.
}
