% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{coherenceToOffset}
\alias{coherenceToOffset}
\title{Convert frequency vs. frequency coherence to offset vs frequency}
\usage{
coherenceToOffset(obj)
}
\arguments{
\item{obj}{An object as returned by \link{coherence}.}
}
\value{
A list containing the centre frequencies, offset values, offset coherences, and 
all the extra information returned by \link{coherence}.
}
\description{
Converts from a "diagonal" coherence matrix to a "horizontal" coherence matrix.
}
\details{
Takes the rows of the diagonal matrix centered at the frequencies 
in \code{freqRange} with \code{maxFreqOffset} values on either side.  These rows then 
become the columns of the returned offset coherence matrix.

The offset is from the frequency in x
}
