% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{coherency.mtm}
\alias{coherency.mtm}
\title{Calculates Frequency Domain Coherence}
\usage{
coherency.mtm(x, y = NULL, xySectioned = FALSE, n = NULL,
  forwardCoh = TRUE, freqRange = NULL, freqOffset = NULL,
  slideFreq = NULL, slideIdx = 1, nw = NULL, k = NULL, deltat = 1,
  nFFT = NULL, nFFT.add = 3, blockSize = NULL, overlap = NULL,
  adaptive = TRUE, jackknife = FALSE, prewhiten = FALSE,
  removePeriodic = TRUE, sigCutoff = NULL, nodes = 1,
  clusterType = c("SOCK", "MPI"))
}
\arguments{
\item{x}{A \code{vector} of values for the first dataset.}

\item{y}{A \code{vector} of values for the second dataset - if null, assumes ... 
I'm not sure actually.}
}
\value{
Returns a matrix containing the coherency.
}
\description{
Estimates and returns the Loeve spectrum given series x and y.
}
