% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{eigenByFreq}
\alias{eigenByFreq}
\title{Creates a design matrix at a particular frequency}
\usage{
eigenByFreq(obj, offsets = NULL, rowNum, numEl)
}
\arguments{
\item{obj}{- a list of lists of matrices - each sublist is a time block.  Within each time 
block list is a matrix for each input with nFFT rows and K columns}

\item{rowNum}{- the current frequency that we're building the design matrix for}

\item{numEl}{- the number of columns in the design matrix without offsets}

\item{offests}{- the result of offsetFreq() (I think.. )}
}
\description{
Design matrix designing! Using offsets if applicable.
}
