% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{freqOffsetAxis}
\alias{freqOffsetAxis}
\title{Frequency axis calculation}
\usage{
freqOffsetAxis(nFFT, freqOffset, deltat)
}
\arguments{
\item{nFFT}{A \code{numeric} giving total number of frequency bins.}

\item{freqOffset}{A \code{numeric} giving the maximum offset used.}

\item{deltat}{A \code{numeric} giving the sampling rate of the time series.}
}
\value{
A \code{vector} of frequency values with range (0, nyquist) inclusive.
}
\description{
Uses number of frequency bins and deltat values to create frequency axis.
}
