% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset.R
\name{freqUsed}
\alias{freqUsed}
\title{Returns the frequencies used to estimate transfer functions}
\usage{
freqUsed(obj)
}
\arguments{
\item{obj}{An object as returned by \code{transfer::offsetFreq}}
}
\description{
Absolute frequencies (and possibly relative (i.e., offsets))
}
