% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{H2zero}
\alias{H2zero}
\title{Sets the low frequency component of a transfer function to zero}
\usage{
H2zero(H, f, freqColName = "freq")
}
\arguments{
\item{H}{Returned from \code{tf()} - assumed \code{asMatrix == TRUE} and a "freq" column exists.}

\item{f}{The last frequency to be zero'd out.}

\item{freqColName}{The name of the frequency column, probably "freq" (default)}
}
\description{
Used when you have high pass filtered data
}
