% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{impulseResponse}
\alias{impulseResponse}
\title{Obtain the impusle response from a transfer object}
\usage{
impulseResponse(object, frequencyName = "freq")
}
\arguments{
\item{object}{An object of class \code{transfer} obtained from \link{tf.svd} or \link{tf}.}
}
\value{
A vector containing the impulse response of the provided transfer function 
from \link{tf}.
}
\description{
Inverse Fourier transforms the transfer function to obtain the impulse response
}
\details{
Inverts the transfer function object returning causal and non-causal filter 
coefficients.  The non-causal filter coefficients are the left "half" of the vector with 
the causal parts being on the right (this is the form required for \link{filter}).
Aaron maybe want to make sure I didn't say this incorrectly O_O.
}
