% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{marginalCoh}
\alias{marginalCoh}
\title{Averages across the rows of the offset coherence matrix}
\usage{
marginalCoh(obj, msc = TRUE)
}
\arguments{
\item{obj}{A \code{list} as returned by \link{coherenceToOffset}.}
}
\description{
Estimates the average coherence for each offset within a frequency band.
}
