% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{offsetFreq}
\alias{offsetFreq}
\title{Determine Offset Frequencies to include in model}
\usage{
offsetFreq(dat, responseName, blockSize = dim(dat)[1], overlap = 0,
  deltat = 1, nw = 4, k = 7, nFFT = NULL, mscCutoff = 0.2,
  freqRange = NULL, maxFreqOffset = 0, calcType = 1, forward = 1)
}
\arguments{
\item{dat}{a \code{data.frame} whose columns contain the time-domain response
and predictors}

\item{responseName}{a \code{character} string giving the name of which column is the 
response.}

\item{blockSize}{The length of a single block to use (if blocking)}

\item{overlap}{A \code{numeric} value in the range [0, 1) indicating the proporation of 
overlap between neighbouring blocks.}

\item{deltat}{The sampling rate in seconds.}

\item{nw}{time-bandwidth parameter for multitaper}

\item{k}{number of tapers to use (k < 2*nw)}

\item{nFFT}{the number of frequency bins to use (nFFT > 2*ndata)}

\item{freqRange}{A vector with 2 elements containing the start and end frequencies (in Hz) 
over which to calculate the coherence.}

\item{maxFreqOffset}{Every pair of frequencies between f1 (series 1) 
and f1 +/- maxFreqOffset (series 2) will be calculated (f1 + maxFreqOffset < nyquist)}

\item{calcType}{An \code{integer} value indicating how averaging over blocks should 
be performed:
1 - calculate the MSC on each block, then average;
2 - calculate the cross and auto spectra on each block, average each quantity 
across blocks, then calculate the coherency;
3 - calculate the coherency on each block, then average
4 - minimum MSC across blocks}

\item{forward}{An \code{integer} indicating whether the forward (1) or reverse (0) coherence
should be calculated.}
}
\description{
Calculates the coherence between multiple inputs and an output and determines 
significant offset frequencies
}
