% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{olsTf}
\alias{olsTf}
\title{Ordinary least squares frequency-domain regression}
\usage{
olsTf(x, y, time, n, npredictor, ntaper, freq, fOffset)
}
\arguments{
\item{x}{is a \code{list} of \code{lists}: each sublist is a block of time containing 
numeric matrices like y - one matrix for each predictor}

\item{y}{a \code{list} of complex matrices (containing real data probably) 
- each matrix is a block in time and contains response * dpss - n rows, k columns}

\item{time}{A vector containing times at which the data are sampled for each block.
(you can just supply 1:n)}

\item{n}{An \code{integer} indicating the number of samples in each block.}

\item{npredictor}{An \code{integer} indicating the number of inputs (X's)}

\item{ntaper}{An \code{integer} indicating the number of tapers that were used.}

\item{freq}{is a list of frequencies at which to estimate the transfer function 
- someone needs to take into account that there will be a gap at beginning and 
end of size max(fOffset).}

\item{fOffset}{a vector indicating which frequencies of predictors to use 
(should at least be 0)}
}
\description{
Estimates coefficients by frequency (transfer function) between 
time-domain inputs \code{x} and response \code{y}.
}
\details{
I need to revisit this and think of a way to NOT use the slow Fourier transform.
Also, probably want to be using an SVD regression to get the transfer function.
}
