% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{olsTfEigen}
\alias{olsTfEigen}
\title{Ordinary least squares frequency-domain regression on eigencoefficients}
\usage{
olsTfEigen(x, y, freqIdx, fOffsetIdxLst)
}
\arguments{
\item{x}{\code{list} of inputs that is a result of \code{taper()}.}

\item{y}{\code{list} of responses that is a result of \code{taper()}.}

\item{freqIdx}{\code{numeric} vector indicating at which frequencies the 
coefficients should be estimated}

\item{fOffsetIdxLst}{a \code{list} containing what offset frequencies should be used 
(in a standard transfer function, this would contain the zero frequency offset)}
}
\value{
a \code{matrix} with columns the transfer function relating to the input 
that was in the same column.
}
\description{
Estimates coefficients by frequency (transfer function) between 
inputs \code{x} and response \code{y}.
}
\details{
Note to me: I should look at exactly how this is working (re: fOffsetIdxLst)
}
