% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset.R
\name{plot.freqUsed}
\alias{plot.freqUsed}
\title{Plots an object as returned from transfer::freqUsed}
\usage{
\method{plot}{freqUsed}(obj, fmf = 1, plot1to1 = TRUE, ...)
}
\arguments{
\item{obj}{What gets returned from transfer::freqUsed function}

\item{fmf}{The Frequency Multiplication Factor (i.e. - 1e6 for uHz labels)}

\item{plot1to1}{A \code{logical} indicating whether to plot the 1 to 1 line 
(this would be the standard frequency pairings used).}
}
\description{
Shows the distribution of points used in the transfer function estimation
}
\details{
Only 5 colours currently - need to deal with this in the future perhaps.
}
