% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_hello_world}
\alias{rcpp_hello_world}
\title{The Title! Hello World}
\usage{
rcpp_hello_world()
}
\arguments{
\item{x}{Something about a parameter}
}
\value{
Nothing.
}
\description{
The description
}
\details{
Some details about this function.
}
