% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sd.complex}
\alias{sd.complex}
\title{Standard deviation of complex values}
\usage{
sd.complex(x)
}
\arguments{
\item{x}{a \code{vector} containing real or complex values.}
}
\value{
A real valued scalar.
}
\description{
Calculates the standard deviation of a complex-valued series.
}
\details{
Calculates the standard deviation of either a real or complex series using
sd = E{(X-u)(X-u)'}
where the apostrophe denotes the complex conjugate.
}
