% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sectionData}
\alias{sectionData}
\title{Section data into data blocks}
\usage{
sectionData(x, blockSize, overlap = 0)
}
\arguments{
\item{x}{A \code{data.frame} where each column is assumed to be a series.}

\item{blockSize}{An \code{integer} indicating the size of each section.}

\item{overlap}{An \code{numeric} between 0 and 1 indicating how much overlap should 
exist between sections.}
}
\value{
A \code{list} whose elements contain a \code{data.frame} with the same number of 
columns as \code{x} and \code{blockSize} number of rows.
}
\description{
Sections provided data using block size and overlap proportion
}
