% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sftCpp}
\alias{sftCpp}
\title{Slow Fourier transform}
\usage{
sftCpp(x, f, time)
}
\arguments{
\item{x}{A complex vector to be Fourier transformed}

\item{f}{A \code{numeric} value indicating the frequency
at which to calculate the Fourier tranform}

\item{time}{A vector containing the times (sampling rate sort of).}
}
\value{
complex value of the Fourier transform of x at f.
}
\description{
Computes the Fourier transfrom of the data \code{x} at the frequency \code{f}.
}
\details{
My understanding is that performing the Fourier transform in this way 
is numerically less stable than using an FFT and associated algorithms.
This is something that should be addressed by the author.
}
