% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon.R
\name{specOffRecon}
\alias{specOffRecon}
\title{Testing the fit with offset coherences}
\usage{
specOffRecon(newData, H, nw, k, nFFT, deltat)
}
\arguments{
\item{newData}{a \code{data.frame} containing the same columns that were used in the 
original transfer function fitting.}

\item{H}{an object of type \code{transfer} from tf().}

\item{nw}{time-bandwidth parameter (needed for window function estimation) - same value
as used to estimate \code{H}.}

\item{k}{an \code{integer} indicating the number of tapers used in estimation of 
\code{H} - although maybe they don't need to be the same for this part...}
}
\description{
Frequency domain exploration of fit.
}
