% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stackEigenByFreq}
\alias{stackEigenByFreq}
\title{Need to write this..}
\usage{
stackEigenByFreq(x, y = NULL)
}
\description{
Should write a description too..
}
\details{
I don't think I need this except for during transfer()...
}
