% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{std}
\alias{std}
\title{Standardize a vector}
\usage{
std(x)
}
\arguments{
\item{x}{A \code{numeric} vector to standardize.}
}
\value{
A standardized vector.
}
\description{
Subtracts the mean and divides by the standard deviation.
}
\details{
Does exactly what the description says and nothing else.  Will also 
work on complex-valued vectors.
}
