% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{taper}
\alias{taper}
\title{Multiplies a vector by a data taper}
\usage{
taper(x, dataTaper = "dpss", nw = 4, k = 7)
}
\arguments{
\item{x}{A \code{list} of \code{data.frame}'s as a result of calling \code{sectionData()}.}

\item{dataTaper}{A \code{character} string indicating what type of taper to use.
Currently only DPSS's tapers are available.}

\item{nw}{A \code{numeric} representing the time-bandwidth paramater for use in generating 
the DPSS's.}

\item{k}{A \code{numeric} indicating the number of tapers to be used.  Should be less 
than 2*nw.}
}
\value{
A \code{list} of \code{lists}, each "sublist" a named matrix of length 
\code{blockSize} by \code{k} of tapered data for that series.
}
\description{
Tapers the data using a discrete prolate spheroidal sequence (DPSS)
}
\details{
This is not implemented in an efficient manner - there needs to be a better way
written.  That will be in the future.
}
