% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{tf}
\alias{tf}
\title{Estimate the frequency-domain transfer function}
\usage{
tf(x, y, blockSize = dim(x)[1], overlap = 0, deltat = 1, nw = 4,
  k = 7, nFFT = NULL, freqRange = NULL, freqOffset = NULL)
}
\arguments{
\item{x}{A \code{data.frame} whose columns are the time domain input series.}

\item{y}{A single column \code{data.frame} containing the response series.}

\item{blockSize}{A \code{numeric} indicating the block sizes into which the 
input and response series will be partitioned.}

\item{overlap}{A \code{numeric} between 0 and 1 indicating how much overlap should exist 
between adjacent blocks.}

\item{deltat}{A \code{numeric} indicating the sample rate.}

\item{nw}{A \code{numeric} indicating the time bandwidth parameter for estimating the 
Slepian data tapers.}

\item{k}{A \code{numeric} indicating the number of tapers to use - should be approximately
floor(2*nw - 1) and no larger than floor(2*nw).}

\item{nFFT}{A \code{numeric} indicating the number of frequency bins to use (i.e. setting 
the zeropadding amount).}

\item{freqRange}{NOT CURRENTLY IMPLEMENTED.}

\item{freqOffset}{NOT CURRENTLY IMPLEMENTED (don't chage this... ).}
}
\value{
A complex matrix whose columns are the individual transfer function for each input.
}
\description{
Estimates a transfer function between the columns of input \code{x} and the response
\code{y}.
}
\details{
Takes the times series inputs and response, divides these series into 
(optionally) overlapping blocks, tapers each block with Discrete 
Prolate Spheriodal Sequences (DPSS's or Slepian sequences), Fourier transforms each 
block, and then estimates the transfer function at each frequency between the Fourier 
transforms of the inputs and the response.
}
