% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{tf_content}
\alias{tf_content}
\title{Get the content}
\usage{
tf_content(file)
}
\arguments{
\item{file}{Access the content of the local file that are connected with the link.}
}
\description{
This function accesses the content of the file.
}
