% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{tf_upload}
\alias{tf_upload}
\title{Upload a file to \code{transfer.sh}}
\usage{
tf_upload(
  filename,
  path = NULL,
  max_downloads = NULL,
  max_days = NULL,
  spinner = TRUE,
  ...
)
}
\arguments{
\item{filename}{File to be uploaded}

\item{path}{Path to find the file (combined with filenamed)}

\item{max_downloads}{Number of max downloads downloaded from the provided link.}

\item{max_days}{Number of days that the link is going be live. Maximum to 14.}

\item{spinner}{Whether to show a reassuring spinner while the process is running.}

\item{...}{Further arguments paseed to \code{\link[processx]{run}}.}
}
\value{
a \code{transfer.sh} link.
}
\description{
Share files with a URL Upload up to 10 GB Files stored for 14 days
}
\examples{
\dontrun{

# Upload current directory
x <- tf_upload(".")

# See the content send to \code{transfer.sh}
tf_content(x)

# Open the link in browser
tf_browse(x)

# Download
tf_download(x)
}
}
