% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_operations.R
\name{transfer_describe_server}
\alias{transfer_describe_server}
\title{Describes the server that you specify by passing the ServerId parameter}
\usage{
transfer_describe_server(ServerId)
}
\arguments{
\item{ServerId}{[required] A system-assigned unique identifier for an SFTP server.}
}
\description{
Describes the server that you specify by passing the \code{ServerId}
parameter.
}
\details{
The response contains a description of the server's properties.
}
\section{Request syntax}{
\preformatted{transfer$describe_server(
  ServerId = "string"
)
}
}

\keyword{internal}
