% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zFilter}
\alias{zFilter}
\title{Filters a time-series}
\usage{
zFilter(x, filter)
}
\arguments{
\item{x}{A vector (possibly complex, time-series) to be filtered.}

\item{filter}{A vector (possibly complex) of filter coefficients to be applied.}
}
\description{
Allows complex-valued filtering
}
\details{
The causal coefficients should be in the right-half of the vector.  This function 
assumes that the filter is odd-length.
}
