% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{coherency.mtm}
\alias{coherency.mtm}
\title{Calculates Frequency Domain Coherence}
\usage{
coherency.mtm(x, y = NULL, xySectioned = FALSE, n = NULL,
  forwardCoh = TRUE, freqRange = NULL, freqOffset = NULL,
  slideFreq = NULL, slideIdx = 1, nw = NULL, k = NULL, deltat = 1,
  nFFT = NULL, nFFT.add = 3, blockSize = NULL, overlap = NULL,
  adaptive = TRUE, jackknife = FALSE, prewhiten = FALSE,
  removePeriodic = TRUE, sigCutoff = NULL, nodes = 1,
  clusterType = c("SOCK", "MPI"))
}
\arguments{
\item{x}{A \code{vector} of values for the first dataset.}

\item{y}{A \code{vector} of values for the second dataset - if null, assumes ... 
I'm not sure actually.}

\item{xySectioned}{A \code{logical} denoting whether the data has already been sectioned.
\code{FLASE} implies that the code will be sectioned according to \code{blockSize} and 
\code{overlap}.}

\item{n}{A \code{numeric} value giving the length of the \code{x} and \code{y} data sets. 
If \code{NULL}, will be determined in the function.}

\item{forwardCoh}{A \code{logical} value indicating if the forward coherence (\code{TRUE}) 
or reverse coherence \code{FALSE} should be computed.}

\item{freqRange}{A \code{vector} of two values containing the range of frequencies to use.
This range should be contained within [0, 1/(2*\code{dt})] (0 to Nyquist)}
}
\value{
Returns a matrix containing the coherency.
}
\description{
Estimates and returns the Loeve spectrum given series x and y.
}
