% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{predict.transfer}
\alias{predict.transfer}
\title{Predict using an estimated frequency-domain transfer function}
\usage{
\method{predict}{transfer}(object, newdata, filterMod = trim, ...)
}
\arguments{
\item{object}{An object of class \code{transfer}, from a call to the \code{\link{tf}} function.}

\item{newdata}{A \code{data.frame} whose columns are the time domain input series.}

\item{filterMod}{A \code{function} to be applied to the filter coefficients before convolution.}

\item{...}{additional arguments passed to \code{filterMod}.}
}
\value{
A \code{data.frame} with the predicted values obtained by filtering 
the input series \code{newdata}.
}
\description{
Using output from the \code{\link{tf}} function, and an input (multivariate) time series,
this function generates predicted values.
}
\details{
The transfer function estimate is used to calculate filter coefficients to be
applied in the time domain via convolution with the input series \code{newdata}.
Prior to the convolution, the filter coefficients can be modified using the
\code{filterMod} function.
}
