% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{tf_download}
\alias{tf_download}
\title{Download from \code{transfer.sh}}
\usage{
tf_download(url = NULL, name = NULL, spinner = TRUE, ...)
}
\arguments{
\item{url}{The url to from.}

\item{name}{Whether to change the name of the download file into a user-specified n
name.}

\item{spinner}{Whether to show a reassuring spinner while the process is running.}

\item{...}{further arguments paseed to \code{\link[processx]{run}}}
}
\description{
Using the link, download the file(s).
}
