% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{tf_upload}
\alias{tf_upload}
\title{Upload a file to \code{transfer.sh}}
\usage{
tf_upload(file, max_downloads = NULL, max_days = NULL,
  spinner = TRUE, ...)
}
\arguments{
\item{file}{the file to be uploaded}

\item{max_downloads}{number of max downloads}

\item{max_days}{number of days that the link is gonna be live. Maximum to 14.}

\item{spinner}{Whether to show a reassuring spinner while the process is running.}

\item{...}{further arguments paseed to \code{\link[processx]{run}}}
}
\description{
Share files with a URL Upload up to 10 GB Files stored for 14 days
}
\examples{
\dontrun{

# Upload current directory
tf_upload(".")
}
}
